+++ spec/hausgold/global_id_spec.rb
@@ -1,11 +1,95 @@
 # frozen_string_literal: true
 
-RSpec.describe Hausgold::Url do
+RSpec.describe Hausgold::GlobalId do
   let(:described_class) { Hausgold }
   let(:uuid) { 'c150681f-c514-438a-8413-7c8f24a5f9dd' }
   let(:other_uuid) { 'c7ee24fc-4ccf-4205-b087-ff711d25ef3f' }
 
-  describe '#locate' do
+  describe 'initialization' do
+    let(:all_locators_count) { Hausgold::Configuration::API_NAMES.count }
+
+    before { allow(GlobalID::Locator).to receive(:use) }
+
+    after { Class.new { include Hausgold::GlobalId } }
+
+    context 'with configured local app exclusion' do
+      before { Hausgold.configuration.exclude_local_app_gid_locator = true }
+
+      context 'with configured local app name' do
+        before { Hausgold.configuration.app_name = 'identity-api' }
+
+        it 'registers not all GID locators' do
+          expect(GlobalID::Locator).to \
+            receive(:use).exactly(all_locators_count - 1).times
+        end
+      end
+
+      context 'without configured local app name (auto detection)' do
+        before do
+          add_fake_rails_app('IdentityApi')
+          # We need to reset the +app_name+ with the configuration defaults,
+          # because the defaults are not evaluated again on reset
+          Hausgold.configuration.app_name = Hausgold.local_app_name
+        end
+
+        after { remove_fake_rails_app('IdentityApi') }
+
+        it 'registers not all GID locators' do
+          expect(GlobalID::Locator).to \
+            receive(:use).exactly(all_locators_count - 1).times
+        end
+      end
+
+      context 'without configured local app name (no auto detection)' do
+        before { Hausgold.configuration.app_name = nil }
+
+        it 'registers all GID locators' do
+          expect(GlobalID::Locator).to \
+            receive(:use).exactly(all_locators_count).times
+        end
+      end
+    end
+
+    context 'without configured local app exclusion' do
+      before { Hausgold.configuration.exclude_local_app_gid_locator = false }
+
+      context 'with configured local app name' do
+        before { Hausgold.configuration.app_name = 'identity-api' }
+
+        it 'registers all GID locators' do
+          expect(GlobalID::Locator).to \
+            receive(:use).exactly(all_locators_count).times
+        end
+      end
+
+      context 'without configured local app name (auto detection)' do
+        before do
+          add_fake_rails_app('IdentityApi')
+          # We need to reset the +app_name+ with the configuration defaults,
+          # because the defaults are not evaluated again on reset
+          Hausgold.configuration.app_name = Hausgold.local_app_name
+        end
+
+        after { remove_fake_rails_app('IdentityApi') }
+
+        it 'registers all GID locators' do
+          expect(GlobalID::Locator).to \
+            receive(:use).exactly(all_locators_count).times
+        end
+      end
+
+      context 'without configured local app name (no auto detection)' do
+        before { Hausgold.configuration.app_name = nil }
+
+        it 'registers all GID locators' do
+          expect(GlobalID::Locator).to \
+            receive(:use).exactly(all_locators_count).times
+        end
+      end
+    end
+  end
+
+  describe '.locate' do
     let(:user_uuid) { 'bf136aed-0259-4458-8cf7-762553eebfc2' }
     let(:gid_raw) { "gid://calendar-api/Task/#{uuid}" }
     let(:gid_raw_user) { "gid://identity-api/User/#{user_uuid}" }
@@ -30,7 +114,7 @@ RSpec.describe Hausgold::Url do
     end
   end
 
-  describe '#build_gid' do
+  describe '.build_gid' do
     let(:app) { 'identity-api' }
     let(:gid) { "gid://#{app}/Task/#{uuid}" }