+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added support to download Asset files (`Hausgold::Asset.find('..').download`)
 * Implemented an initial type-casted attribute system (writers and readers)
   and added it to all entities (booleans, symbols, string inquirers)
+* Added an automatic `app_name` configuration guessing (Rails apps only)
+* Added a configuration (`exclude_local_app_gid_locator`) to toggle the
+  exclusion of same named local application and Hausgold GID locators
 
 ### 0.3.0